fix: created individual  component
